Project/ppssfl_project/requirements.txt
flwr
torch
numpy
pandas
matplotlib
scikit-learn
opacus
streamlitplotly